products update

+++ backend/helpers/fileHelpers.js
@@ -14,13 +14,14 @@ const addFile = async (file, folderName) => {
 };
 
 const addMultipleFiles = async (files, folderName) => {
+  if (!files) return;
+
   const folderPath = path.join(__dirname, `../uploads/${folderName}`);
   const baseURL = `${process.env.BASE_URL}/uploads/${folderName}`;
 
   const imagesURL = [];
 
   for (const file of files) {
-    console.log("file", file);
     const uniqueName = Date.now() + "-" + file.name;
     await file.mv(path.join(folderPath, uniqueName));
     imagesURL.push(`${baseURL}/${uniqueName}`);
@@ -39,14 +40,25 @@ const deleteFile = async (fileURL, folderName) => {
   }
 };
 
-const deleteMultipleFiles = async (filesURL, folderName) => {
+const deleteMultipleFiles = async (filesURLs, folderName, bodyURLs) => {
   const folderPath = path.join(__dirname, `../uploads/${folderName}`);
   const filesInFolder = await fs.readdir(folderPath);
 
-  for (const url of filesURL) {
-    const fileName = path.parse(url).base;
-    if (filesInFolder.includes(fileName)) {
-      await fs.unlink(path.join(folderPath, fileName));
+  if (bodyURLs) {
+    for (const url of filesURLs) {
+      if (!bodyURLs.includes(url)) {
+        const fileName = path.parse(url).base;
+        if (filesInFolder.includes(fileName)) {
+          await fs.unlink(path.join(folderPath, fileName));
+        }
+      }
+    }
+  } else {
+    for (const url of filesURLs) {
+      const fileName = path.parse(url).base;
+      if (filesInFolder.includes(fileName)) {
+        await fs.unlink(path.join(folderPath, fileName));
+      }
     }
   }
 };
